requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pytz==2018.7
